requirements.txt
streamlit
PyMuPDF
spacy
rapidfuzz
python-docx==1.1.2
pillow
numpy